fixed issue, testing to see same output

+++ proof_of_concept/pipeline.c
@@ -45,7 +45,7 @@ int main (void)
 
     int c_size = 8;
     int num_chunks = (dim_y) / (c_size);
-    float chunks[num_chunks][8][8];
+    int chunks[num_chunks][8][8];
     float result_blks[num_chunks][8][8];
     
     // int n = 0;
